+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Event Admin Required to Fill Out Survey [#3581](https://github.com/rokwire/illinois-app/issues/3581).
 
 ## [5.0.56] - 2023-08-28
 ### Added